meta-mender-tegra: add support for jetson-nano development kit
Jetson-nano development kit has a SD-card slot and no emmc. It does not support
nvidia redundant boot feature and therefore there is no dependency to the BUP Payload which
is used on jetson-tx2. Nvidia released a production module of the jetson-nano with 16GB emmc,
but this is not supported by this branch.

Even though there is no emmc, the sd-card is referenced as mmc device in u-boot, which is
why we use ENVIRONMENT_IS_IN_MMC.

Changelog: Title

Signed-off-by: Moritz Marquardt <moritz.marquardt@zeiss.com>
Signed-off-by: Etienne Cordonnier <etienne.cordonnier@taurob.com>